requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.8.0
blinker==1.8.2
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
firecrawl==1.2.4
firecrawl-py==1.2.4
flake8==7.1.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.1
idna==3.7
importlib_metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.0
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.11.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
openai==1.29.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pocket==0.3.6
pre-commit==3.8.0
pycodestyle==2.12.1
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlalchemy-stubs==0.4
tomli==2.0.1
tqdm==4.66.4
types-requests==2.32.0.20240914
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.26.5
websockets==13.0.1
Werkzeug==3.0.3
zipp==3.20.2
